% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beech}
\alias{beech}
\title{Thermogravimetric analysis of beech tree data}
\format{
## `beech`
A data frame with 1544 rows and 4 columns:
\describe{
  \item{time}{Numeric, time in minutes}
  \item{temp}{Numeric, temperature in degrees C}
  \item{mass_loss}{Numeric, mass loss compared to an initial mass of 10.64}
  \item{stage}{Integer, stage of thermogravimetric analysis}
}
}
\source{
<>
}
\usage{
beech
}
\description{
TGA of beech tree sample executed over a 6 stage temperature program. Stage 1: drying phase with dynamic heating of the sample (10C/min) to a drying temperature of 105C under a nitrogen atmosphere. Stage 2: isothermal drying of the sample for 120 minutes at 105C under a nitrogen atmosphere. Stage 3: dynamic heating (pyrolysis phase) of the sample (10C/min) to 750C under a nitrogen atmosphere. Stage 4: Dynamic cooling of the sample (10C/min) to 600C under a nitrogen atmosphere. Stage 5: dynamic heating (combustion phase) of the sample (2.3C/min) to 815C in air. Stage 6: Isothermal combustion (ashing phase) of the sample for 60 minutes to 815C in air.
}
\keyword{datasets}
