% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bounds.R
\name{Bounds}
\alias{Bounds}
\title{Accessor function to extract selected bounds}
\usage{
Bounds(object)
}
\arguments{
\item{object}{the output of either the process or deconvolve functions}
}
\value{
Temperature bounds of the data in the object
}
\description{
Accessor function to extract selected bounds
}
\examples{
data(juncus)
tmp <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
Bounds(tmp)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{temperature}
\keyword{thermogravimetry}
