% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moisture_content.R
\name{calculate_ash_content}
\alias{calculate_ash_content}
\title{Calculate ash content as percent of dry mass of sample}
\usage{
calculate_ash_content(processed_data)
}
\arguments{
\item{processed_data}{thermogravimetric mass loss data processed by
process() function}
}
\value{
ash content %
}
\description{
Calculate ash content as percent of dry mass of sample
}
