% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculative_fixed_carbon_fractions.R
\name{calculate_fixed_carbon_fractions}
\alias{calculate_fixed_carbon_fractions}
\title{Calculate fixed carbon fractions}
\usage{
calculate_fixed_carbon_fractions(
  deconvolution_output,
  dry_basis_fixed_carbon_hemicellulose = 13.8,
  dry_basis_fixed_carbon_cellulose = 6.4,
  dry_basis_fixed_carbon_lignin = 34
)
}
\arguments{
\item{deconvolution_output}{output from the pyrolysis phase deconvolution}

\item{dry_basis_fixed_carbon_hemicellulose}{reference value for proportion
of pure hemicellulose sample that is fixed carbon}

\item{dry_basis_fixed_carbon_cellulose}{reference value for proportion
of pure cellulose sample that is fixed carbon}

\item{dry_basis_fixed_carbon_lignin}{reference value for proportion
of pure lignin sample that is fixed carbon}
}
\value{
list of volatile hemicellulose fraction, volatile cellulose
  fraction, volatile lignin fraction, fixed carbon hemicellulose fraction,
  fixed carbon cellulose fraction, and fixed carbon lignin fraction
}
\description{
Calculate fixed carbon fractions
}
