% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moisture_content.R
\name{calculate_moisture_content}
\alias{calculate_moisture_content}
\title{Calculate moisture content as percent of total sample mass}
\usage{
calculate_moisture_content(processed_data)
}
\arguments{
\item{processed_data}{thermogravimetric mass loss data processed by
process() function}
}
\value{
moisture content %
}
\description{
Calculate moisture content as percent of total sample mass
}
