% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_weights.R
\name{component_weights}
\alias{component_weights}
\title{Accessor function to extract mean weights}
\usage{
component_weights(object)
}
\arguments{
\item{object}{a decon object}
}
\value{
Extract mean fractions of the object
}
\description{
Accessor function to extract mean weights
}
\examples{
\dontrun{
data(juncus)
tmp <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
output <- deconvolve(tmp)
component_weights(output)
}
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
