% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyperus-data.R
\docType{data}
\name{cyperus}
\alias{cyperus}
\title{Thermogravimetric data for Cyperus eragrostis}
\format{An object of class \code{'cross'}}
\source{
Saras M Windecker
}
\usage{
data(cyperus)
}
\description{
Raw thermogravimetric data from the wetland sedge, C. eragrostis.
}
\examples{
data(cyperus)
}
\keyword{datasets}
