% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_plot.R
\name{decon_plot}
\alias{decon_plot}
\title{Plot pseudo-components}
\usage{
decon_plot(temp, deriv, fit, lower, upper)
}
\arguments{
\item{temp}{vector of temperature data in K}

\item{deriv}{vector of DTG data}

\item{fit}{model fit output from deconvolve function}

\item{lower}{lower bound for temperature for cropping dataset}

\item{upper}{upper bound for temperature for cropping dataset}
}
\value{
plot of DTG against K, with curve for each pseudo-component included
}
\description{
This function produces a plot of the overal DTG curve as well as the deconvolved pseudo-components.
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
