% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_function.R
\name{fs_function}
\alias{fs_function}
\title{Fraser-Suzuki function for a single curve}
\usage{
fs_function(temp, height, skew, position, width)
}
\arguments{
\item{temp}{temperature values}

\item{height}{height value}

\item{skew}{shape value}

\item{position}{position value}

\item{width}{width value}
}
\value{
Fraser-Suzuki function
}
\description{
This function calculates the Fraser-Suzuki function.
}
