% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_mixture.R
\name{fs_mixture}
\alias{fs_mixture}
\title{Fraser-Suzuki three-part mixture}
\usage{
fs_mixture(x, params)
}
\arguments{
\item{x}{temperature values}

\item{params}{parameter starting values}
}
\value{
Fraser-Suzuki mixture model output
}
\description{
This function combines three Fraser-Suzuki functions for the total mixture model.
}
\keyword{internal}
