% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_mixture_wrap.R
\name{fs_mixture_wrap}
\alias{fs_mixture_wrap}
\title{Wrapper for Fraser-Suzuki mixture model}
\usage{
fs_mixture_wrap(x, h1, h2, h3, s1, s2, s3, p1, p2, p3, w1, w2, w3)
}
\arguments{
\item{x}{temperature values}

\item{h1}{height value pc1}

\item{h2}{height value pc2}

\item{h3}{height value pc3}

\item{s1}{shape value pc1}

\item{s2}{shape value pc2}

\item{s3}{shape value pc3}

\item{p1}{position value pc1}

\item{p2}{position value pc2}

\item{p3}{position value pc3}

\item{w1}{width value pc1}

\item{w2}{width value pc2}

\item{w3}{width value pc3}
}
\value{
three-part Fraser-Suzuki mixture model output
}
\description{
This function is a wrapper for the fsTotal function that modifies so it uses all 12 parameters as input.
}
\keyword{internal}
