% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_quantiles.R
\name{get_weights}
\alias{get_weights}
\title{Calculate weight quantiles}
\usage{
get_weights(param_vec, output)
}
\arguments{
\item{param_vec}{parameter estimates from minpack model}

\item{output}{deconvolve output of model}
}
\value{
weights for each component
}
\description{
Calculate weight quantiles
}
