% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{juncus}
\alias{juncus}
\title{Thermogravimetric analysis of juncus reed data}
\format{
## `juncus`
A data frame with 46080 rows and 2 columns:
\describe{
  \item{temp_C}{Numeric, temperature in degrees C}
  \item{mass_loss}{Numeric, mass loss compared to an initial mass of 18.96}
  \item{time}{Numeric, time in minutes}
}
}
\source{
<>
}
\usage{
juncus
}
\description{
TGA of Juncus amabilis reed sample under a single a dynamic heating protocol (10C/min) from 31C to 800C under a nitrogen atmosphere.
}
\keyword{datasets}
