% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/marsilea-data.R
\docType{data}
\name{marsilea}
\alias{marsilea}
\title{Thermogravimetric analysis of marsilea herb data}
\format{
## `marsilea`
A data frame with 46080 rows and 2 columns:
\describe{
  \item{temp_C}{Numeric, temperature in degrees C}
  \item{mass_loss}{Numeric, mass loss compared to an initial mass of 15.29}
}

An object of class \code{'cross'}
}
\source{
<>

Saras M Windecker
}
\usage{
marsilea

data(marsilea)
}
\description{
TGA of marsilea herb sample under a single a dynamic heating protocol (10C/min) from XXXX to 750C under a nitrogen atmosphere.

Raw thermogravimetric data from the wetland forb, M. drumondii.
}
\examples{
data(marsilea)
}
\keyword{datasets}
