% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_fractions.R
\name{mass_fractions}
\alias{mass_fractions}
\title{Calculate mg/mg value}
\usage{
mass_fractions(fit, lower, upper, mass_init, mass_loss_vec)
}
\arguments{
\item{fit}{model fit output from deconvolve function}

\item{lower}{lower bound for temperature for cropping dataset}

\item{upper}{upper bound for temperature for cropping dataset}

\item{mass_init}{numeric value of mg of initial sample}

\item{mass_loss_vec}{vector of mass loss values from TGA}
}
\value{
list of mg pseudo-component / mg sample for pseudo-components
1, 2, and 3
}
\description{
This function produces a list of the final mg pseudo-component per
mg of sample for each of the three components.
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
