% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassFrac.R
\name{MassFrac}
\alias{MassFrac}
\title{Accessor function to extract mass fractions}
\usage{
MassFrac(object)
}
\arguments{
\item{object}{a decon object}
}
\value{
$mass_fractions of the object
}
\description{
Accessor function to extract mass fractions
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
output <- deconvolve(munge)
MassFrac(output)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
