% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Accessor function to extract model fit}
\usage{
Model(object)
}
\arguments{
\item{object}{a decon object}
}
\value{
$minpack.lm of the object
}
\description{
Accessor function to extract model fit
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
output <- deconvolve(munge)
Model(output)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
