% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{model_parameters}
\alias{model_parameters}
\title{Accessor function to extract model parameters}
\usage{
model_parameters(object)
}
\arguments{
\item{object}{a decon object}
}
\value{
model parameters from minpack.lm::nlsLM fit
}
\description{
Accessor function to extract model parameters
}
\examples{
\donttest{
data(juncus)
tmp <- process(juncus, init_mass = 18.96,
               temp = 'temp_C', mass_loss = 'mass_loss')
output <- deconvolve(tmp)
model_parameters(output)
}
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
