% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{This function evaluates a model using rmse.}
\usage{
objective(theta, model, x, obs)
}
\arguments{
\item{theta}{values for each of the parameters}

\item{model}{mathematical function to be used to model data}

\item{x}{temperature values}

\item{obs}{DTG mass values}
}
\value{
root mean square error of model output
}
\description{
This function evaluates a model using rmse.
}
\keyword{internal}
