% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.process.R
\name{plot.process}
\alias{plot.process}
\title{Default S3 plot method for process objects (derived from `process()`)}
\usage{
\method{plot}{process}(x, plot_type = NULL, cex = 1, colour_segments = FALSE, ...)
}
\arguments{
\item{x}{process object as generated by process}

\item{plot_type}{defaults to both plots. Can specify 'mass' or 'rate'}

\item{cex}{size of plots features}

\item{colour_segments}{logical. FALSE = Black main curve, no legend;
TRUE = Brackets + labels for stages}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
Default S3 plot method for process objects (derived from `process()`)
}
