% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dtg_quantile.R
\name{plot_dtg_quantile}
\alias{plot_dtg_quantile}
\title{Plot DTG curve at a chosen quantile (or mean) with component curves}
\usage{
plot_dtg_quantile(
  decon_obj,
  prob = 0.5,
  n_draws = 500,
  bw = FALSE,
  seed = NULL
)
}
\arguments{
\item{decon_obj}{deconvolve() output}

\item{prob}{Quantile, values such as 0.5/0.025/0.975; when set to “mean,” estimated using fitted points}

\item{n_draws}{Number of samples}

\item{bw}{TRUE Black and white, FALSE Color}

\item{seed}{Random seed}
}
\description{
Plot DTG curve at a chosen quantile (or mean) with component curves
}
