% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.decon.R
\name{print.decon}
\alias{print.decon}
\title{Default S3 print method for decon object (derived from `deconvolve()`)}
\usage{
\method{print}{decon}(x, ...)
}
\arguments{
\item{x}{decon object as generated by deconvolve}

\item{...}{other options passed to plot}
}
\value{
print output
}
\description{
This function sets up the default print method
for outputs from deconvolve function
}
