% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.process.R
\name{print.process}
\alias{print.process}
\title{Sets default print method for process object}
\usage{
\method{print}{process}(x, ...)
}
\arguments{
\item{x}{process object as generated by deconvolve}

\item{...}{other options passed to plot}
}
\value{
print output
}
\description{
This function sets up the default print method
for outputs from process function
}
