% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Calculates the derivative rate of mass loss of thermogravimetric data}
\usage{
process(
  data,
  init_mass,
  temp,
  mass_loss,
  time,
  pyrolysis_start_time,
  pyrolysis_end_time,
  temp_units = c("C", "K")
)
}
\arguments{
\item{data}{dataframe with thermogravimetric analysis output}

\item{init_mass}{numeric value of initial sample mass in mg}

\item{temp}{column name containing temperature values}

\item{mass_loss}{column name containing mass loss values in mg}

\item{time}{index column of time}

\item{pyrolysis_start_time}{value from `time` column that indicates when
the pyrolysis period begins}

\item{pyrolysis_end_time}{value from `time` column that indicates when
the pyrolysis period ends}

\item{temp_units}{specify units of temperature, default = Celsius.
Can specify 'K' or 'Kelvin' if in Kelvin}
}
\value{
process list containing modified dataframe, initial mass
of sample, and maximum and minimum temperature values
}
\description{
This function processes thermogravimetric data by calculating
the derivative of mass loss
}
\examples{
data(juncus)
tmp <- process(juncus, init_mass = 18.96,
               temp = 'temp_C', mass_loss = 'mass_loss'
               time,
               )

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
