% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Thermogravimetric data for Juncus usitatus}
\format{An object of class \code{"cross"}}
\source{
Saras M Windecker
}
\usage{
data(raw_data)
}
\description{
Raw thermogravimetric data from the wetland rush, J. usitatus.
}
\keyword{datasets}
