% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_select.R
\name{rmse}
\alias{rmse}
\title{Calculate room mean squared error}
\usage{
rmse(pred, obs)
}
\arguments{
\item{pred}{predicted parameter value}

\item{obs}{DTG mass values}
}
\value{
root mean squared error
}
\description{
This function calculated root mean squarer error of prediction.
}
\keyword{internal}
