% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_param.R
\name{single_param}
\alias{single_param}
\title{Calculate single parameter estimates}
\usage{
single_param(fit, param, component)
}
\arguments{
\item{fit}{model fit output from deconvolve function}

\item{param}{parameter value of interest}

\item{component}{Number of pseudo-component of interest}
}
\value{
parameter estimate value
}
\description{
This function gives single parameter estimates for given
parameter and pseudo-component.
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
