% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_summary.R
\name{stage_summary}
\alias{stage_summary}
\title{Summarise stage start/end time and mass}
\usage{
stage_summary(process_obj)
}
\arguments{
\item{process_obj}{object returned from [process()]}
}
\value{
data.frame with stage, time_start/time_end, mass_start/mass_end
}
\description{
Summarise stage start/end time and mass
}
