% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_deriv.R
\name{tga_deriv}
\alias{tga_deriv}
\title{Prepare raw thermogram data}
\usage{
tga_deriv(data, temp, mass_loss, lower, upper)
}
\arguments{
\item{data}{dataframe}

\item{temp}{dataframe column name containing temperature in Kelvin}

\item{mass_loss}{dataframe column name containing mass loss data}

\item{lower}{lower bound for temperature for cropping dataset}

\item{upper}{upper bound for temperature for cropping dataset}
}
\value{
new dataframe
}
\description{
This function prepares raw thermogravimetry data for deconvolution
by rounding temperature values, taking the derivative, and
clipping at upper and lower bounds.
}
\keyword{derivative}
