% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflection.R
\name{three_peaks}
\alias{three_peaks}
\title{Function to convert inflection point to logical argument for presence of fourth peak}
\usage{
three_peaks(x)
}
\arguments{
\item{x}{numeric}
}
\description{
Function to convert inflection point to logical argument for presence of fourth peak
}
\keyword{internal}
