% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_quantiles.R
\name{weight_quantiles}
\alias{weight_quantiles}
\title{Calculate weight quantiles}
\usage{
weight_quantiles(output)
}
\arguments{
\item{output}{dataframe}
}
\value{
list of means and confidence intervals of weight estimates
}
\description{
Calculate weight quantiles
}
