% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{Weights}
\alias{Weights}
\title{Accessor function to extract mean weights}
\usage{
Weights(object)
}
\arguments{
\item{object}{a decon object}
}
\value{
Extract mean fractions of the object
}
\description{
Accessor function to extract mean weights
}
\examples{
data(juncus)
tmp <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
output <- deconvolve(tmp)
Weights(output)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
