% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_component.R
\name{wt_component}
\alias{wt_component}
\title{Calculate weight single component}
\usage{
wt_component(j, fxn, param_vec, lower_temp, upper_temp)
}
\arguments{
\item{j}{component number}

\item{fxn}{function to integrate over}
}
\value{
weight of component
}
\description{
Calculate weight single component
}
