% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bounds.R
\name{Bounds}
\alias{Bounds}
\title{Accessor function to extract selected bounds}
\usage{
Bounds(object)
}
\arguments{
\item{object}{a process object}
}
\value{
$bounds of the object
}
\description{
Accessor function to extract selected bounds
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
Bounds(munge)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
