% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolves Thermogravimetric Data}
\usage{
deconvolve(process_object, n_curves = NULL, lower_temp = 120,
  upper_temp = 700, start_vec = NULL, lower_vec = NULL,
  upper_vec = NULL)
}
\arguments{
\item{process_object}{process object obtained from process function}

\item{n_curves}{number of curves}

\item{lower_temp}{lower temperature bound to crop dataset}

\item{upper_temp}{upper temperature bound to crop dataset}

\item{start_vec}{vector of starting values for nls function. Only specify this vector if
you have selected the number of curves in the n_curves parameter.}

\item{lower_vec}{vector of lower bound values for nls. Only specify this vector if
you have selected the number of curves in the n_curves parameter.}

\item{upper_vec}{vector of upper bound values for nls. Only specify this vector if
you have selected the number of curves in the n_curves parameter.}
}
\value{
decon list containing amended dataframe, bounds, model output, mass fractions
}
\description{
This function deconvolves thermogravimetric data using a three-part
Fraser-Suzuki mixture model
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 18.96)
output <- deconvolve(munge)

data(marsilea)
munge <- process(marsilea, 'temp_C', 'mass_loss', 10.92)
output <- deconvolve(munge)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
