% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolves Thermogravimetric Data}
\usage{
deconvolve(process_object, lower_temp = 120, upper_temp = 700,
  ranseed = 1, n_curves = NULL, start_vec = NULL, lower_vec = NULL,
  upper_vec = NULL)
}
\arguments{
\item{process_object}{process object obtained from process function}

\item{lower_temp}{lower temperature bound to crop dataset, default to 120}

\item{upper_temp}{upper temperature bound to crop dataset, default to 700}

\item{ranseed}{random seed for nloptr optimiser}

\item{n_curves}{number of curves optional specification}

\item{start_vec}{vector of starting values for nls function. Only specify this vector if
you have selected the number of curves in the n_curves parameter.}

\item{lower_vec}{vector of lower bound values for nls. Only specify this vector if
you have selected the number of curves in the n_curves parameter.}

\item{upper_vec}{vector of upper bound values for nls. Only specify this vector if
you have selected the number of curves in the n_curves parameter.}
}
\value{
decon list containing amended dataframe, bounds, model output, mass fractions
}
\description{
This function deconvolves thermogravimetric data using a Fraser-Suzuki mixture model
}
\examples{
data(juncus)
tmp <- process(juncus, 'temp_C', 'mass_loss', 18.96)
output <- deconvolve(tmp)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
