% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolves Thermogravimetric Data}
\usage{
deconvolve(process_object, start_height = "height", start_skew = "skew",
  start_position = "position", start_width = "width")
}
\arguments{
\item{process_object}{process object obtained from process function}

\item{start_height}{vector of height starting values for parameter optimisation in order of curve}

\item{start_skew}{vector of skew starting values for parameter optimisation in order of curve}

\item{start_position}{vector of position starting values for parameter optimisation in order of curve}

\item{start_width}{vector of width starting values for parameter optimisation in order of curve}
}
\value{
decon list containing amended dataframe, bounds, model output, mass fractions
}
\description{
This function deconvolves thermogravimetric data using a three-part
Fraser-Suzuki mixture model
}
\examples{
data(juncus)
output <- deconvolve(munge, c(0.015, 0.013, 0.01), c(-0.15, -0.15, -0.15),
c(540, 600, 700), c(50, 30, 200))

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
