% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolves Thermogravimetric Data}
\usage{
deconvolve(process_object, n_curves = 3, lower = 400, upper = 900)
}
\arguments{
\item{process_object}{process object obtained from process function}

\item{n_curves}{number of curves}

\item{lower}{lower temperature bound to crop dataset}

\item{upper}{upper temperature bound to crop dataset}
}
\value{
decon list containing amended dataframe, bounds, model output, mass fractions
}
\description{
This function deconvolves thermogravimetric data using a three-part
Fraser-Suzuki mixture model
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 16.85, 'C')
output <- deconvolve(munge, n_curves = 3, lower = 420, upper = 860)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
