% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolves Thermogravimetric Data}
\usage{
deconvolve(process_object, lower = 400, upper = 900, n_curves = NULL)
}
\arguments{
\item{process_object}{process object obtained from process function}

\item{lower}{lower temperature bound to crop dataset}

\item{upper}{upper temperature bound to crop dataset}

\item{n_curves}{number of curves}
}
\value{
decon list containing amended dataframe, bounds, model output, mass fractions
}
\description{
This function deconvolves thermogravimetric data using a three-part
Fraser-Suzuki mixture model
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 16.85, 'C')
output <- deconvolve(munge, lower = 420, upper = 860)

data(marsilea)
munge <- process(marsilea, 'temp_C', 'mass_loss', 10.92, 'C')
output <- deconvolve(munge)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
