% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolves Thermogravimetric Data}
\usage{
deconvolve(data, temp, obs)
}
\arguments{
\item{data}{dataframe}

\item{temp}{temperature values in Kelvin}

\item{obs}{Derivative values of mass loss/K}
}
\value{
model output
}
\description{
This function deconvolves thermogravimetric data using a three-part
Fraser-Suzuki mixture model
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
