% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Optimise Parameters}
\usage{
deconvolve(data, temp, obs)
}
\arguments{
\item{data}{vector of starting values for each of 12 parameters}

\item{temp}{temperature values}

\item{obs}{DTG mass values}
}
\value{
model output
}
\description{
This function otimises parameter selection so that the model will ultimately converge.
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
