% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_function.R
\name{fs_function}
\alias{fs_function}
\title{Fraser-Suzuki function}
\usage{
fs_function(x, h, s, p, w)
}
\arguments{
\item{x}{temperature values}

\item{h}{height value}

\item{s}{shape value}

\item{p}{position value}

\item{w}{width value}
}
\value{
Fraser-Suzuki model output
}
\description{
This function calculates the Fraser-Suzuki function.
}
\keyword{internal}
