% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_mixture.R
\name{fs_mixture}
\alias{fs_mixture}
\title{Mixture model for TGA data}
\usage{
fs_mixture(x, h1, s1, p1, w1, h2, s2, p2, w2, h3, s3, p3, w3, h4 = NULL,
  s4 = NULL, p4 = NULL, w4 = NULL)
}
\arguments{
\item{x}{temperature values}

\item{h1}{height value for curve 1}

\item{s1}{shape value for curve 1}

\item{p1}{position value for curve 1}

\item{w1}{width value for curve 1}

\item{h2}{height value for curve 2}

\item{s2}{shape value for curve 2}

\item{p2}{position value for curve 2}

\item{w2}{width value for curve 2}

\item{h3}{height value for curve 3}

\item{s3}{shape value for curve 3}

\item{p3}{position value for curve 3}

\item{w3}{width value for curve 3}

\item{h4}{height value for curve 4, if present}

\item{s4}{shape value for curve 4, if present}

\item{p4}{position value for curve 4, if present}

\item{w4}{width value for curve 4, if present}
}
\value{
Fraser-Suzuki model output
}
\description{
Mixture model for TGA data
}
