% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_mixture.R
\name{fs_mixture}
\alias{fs_mixture}
\title{Mixture model for TGA data}
\usage{
fs_mixture(x, h1, s1, p1, w1, h2, s2, p2, w2, h3, s3, p3, w3, h0 = NULL,
  s0 = NULL, p0 = NULL, w0 = NULL)
}
\arguments{
\item{x}{temperature values}

\item{h1}{height value for curve 1}

\item{s1}{shape value for curve 1}

\item{p1}{position value for curve 1}

\item{w1}{width value for curve 1}

\item{h2}{height value for curve 2}

\item{s2}{shape value for curve 2}

\item{p2}{position value for curve 2}

\item{w2}{width value for curve 2}

\item{h3}{height value for curve 3}

\item{s3}{shape value for curve 3}

\item{p3}{position value for curve 3}

\item{w3}{width value for curve 3}

\item{h0}{height value for curve 0, if present}

\item{s0}{shape value for curve 0, if present}

\item{p0}{position value for curve 0, if present}

\item{w0}{width value for curve 0, if present}
}
\value{
Fraser-Suzuki model output
}
\description{
Mixture model for TGA data
}
