% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflection.R
\name{inflection}
\alias{inflection}
\title{Function to determine x-value of peak < 500 K}
\usage{
inflection(x, y, w = 1, ...)
}
\arguments{
\item{x}{x-values}

\item{y}{y-values}
}
\value{
returns a list of values. $x returns the inflections points within x.
}
\description{
Function to determine x-value of peak < 500 K
}
\keyword{internal}
