% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/juncus-data.R
\docType{data}
\name{juncus}
\alias{juncus}
\title{Thermogravimetric analysis of juncus reed data}
\format{
## `juncus`
A data frame with 46080 rows and 2 columns:
\describe{
  \item{temp_C}{Numeric, temperature in degrees C}
  \item{mass_loss}{Numeric, mass loss compared to an initial mass of 18.96}
}

An object of class \code{'cross'}
}
\source{
<>

Saras M Windecker
}
\usage{
juncus

data(juncus)
}
\description{
TGA of juncus reed sample under a single a dynamic heating protocol (10C/min) from XXXX to 750C under a nitrogen atmosphere.

Raw thermogravimetric data from the wetland rush, J. amabilis
}
\examples{
data(juncus)
}
\keyword{datasets}
