% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juncus-data.R
\docType{data}
\name{juncus}
\alias{juncus}
\title{Thermogravimetric data for Juncus amabilis}
\format{An object of class \code{'cross'}}
\source{
Saras M Windecker
}
\usage{
data(juncus)
}
\description{
Raw thermogravimetric data from the wetland rush, J. amabilis
}
\examples{
data(juncus)
}
\keyword{datasets}
