% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marsilea}
\alias{marsilea}
\title{Thermogravimetric analysis of marsilea herb data}
\format{
## `marsilea`
A data frame with 46080 rows and 2 columns:
\describe{
  \item{temp_C}{Numeric, temperature in degrees C}
  \item{mass_loss}{Numeric, mass loss compared to an initial mass of 15.29}
  \item{time}{Numeric, time in minutes}
}
}
\source{
<>
}
\usage{
marsilea
}
\description{
TGA of wetland forb, Marsilea drumondii, sample under a single a dynamic heating protocol (10C/min) from 31C to 800C under a nitrogen atmosphere.
}
\keyword{datasets}
