% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marsilea-data.R
\docType{data}
\name{marsilea}
\alias{marsilea}
\title{Thermogravimetric data for Marsilea drumondii}
\format{An object of class \code{'cross'}}
\source{
Saras M Windecker
}
\usage{
data(marsilea)
}
\description{
Raw thermogravimetric data from the wetland forb, M. drumondii.
}
\examples{
data(marsilea)
}
\keyword{datasets}
