% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModData.R
\name{ModData}
\alias{ModData}
\title{Accessor function to extract processed dataframe}
\usage{
ModData(object)
}
\arguments{
\item{object}{a process object}
}
\value{
$data of the object
}
\description{
Accessor function to extract processed dataframe
}
\examples{
data(juncus)
munge <- process(juncus, 'temp_C', 'mass_loss', 18.96, temp_type = 'C')
ModData(munge)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
