% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Accessor function to extract model fit}
\usage{
model_fit(object)
}
\arguments{
\item{object}{a decon object}
}
\value{
$minpack.lm of the object
}
\description{
Accessor function to extract model fit
}
\examples{
\donttest{
data(juncus)
tmp <- process(juncus, init_mass = 18.96,
               temp = 'temp_C', mass_loss = 'mass_loss')
output <- deconvolve(tmp)
model_fit(output)
}
}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
