% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_select.R
\name{objective}
\alias{objective}
\title{This function evaluates root mean squared error of a model}
\usage{
objective(theta, model, temp, obs)
}
\arguments{
\item{theta}{values for each of the parameters}

\item{model}{mathematical function to be used to model data}

\item{temp}{temperature values}

\item{obs}{DTG mass values}
}
\value{
root mean square error of model output
}
\description{
This function evaluates root mean squared error of a model
}
\keyword{internal}
