% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_select.R
\name{param_select}
\alias{param_select}
\title{Optimise parameters}
\usage{
param_select(theta, lb, ub, model, x, obs, restarts = 300, ranseed = 2718)
}
\arguments{
\item{theta}{vector of starting values for each of 12 parameters}

\item{lb}{vector of lower bounds on each of 12 parameters}

\item{model}{mathematical function to be applied}

\item{x}{temperature values}

\item{obs}{DTG mass values}

\item{restarts}{number of times for optimiser to restart}

\item{ranseed}{random seed for optimiser}
}
\value{
optimised starting parameter values
}
\description{
This function otimises parameter selection so that the model will ultimately converge.
}
\keyword{internal}
