% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_select.R
\name{param_select}
\alias{param_select}
\title{Optimise parameters}
\usage{
param_select(theta, lb, ub, model, temp, obs, seed, restarts = 300)
}
\arguments{
\item{theta}{vector of starting values for each parameter}

\item{lb}{vector of lower bounds on each parameter}

\item{ub}{vector of upper bounds on each parameter}

\item{model}{mathematical function to be applied}

\item{temp}{temperature values}

\item{obs}{DTG mass values}

\item{seed}{random seed for optimiser}

\item{restarts}{number of times for optimiser to restart}
}
\value{
optimised starting parameter values
}
\description{
This function otimises parameter selection so that the model will ultimately converge.
}
\keyword{internal}
