% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.decon.R
\name{plot.decon}
\alias{plot.decon}
\title{Sets default plotting for decon object}
\usage{
\method{plot}{decon}(x, plot_type = NULL, ...)
}
\arguments{
\item{x}{decon object as generated by deconvolve}

\item{plot_type}{specify if want the basic graph with 'raw', else defaults to the mixture graph}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
This function sets up the default plotting method
for outputs from deconvolve function
}
