% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.decon.R
\name{plot.decon}
\alias{plot.decon}
\title{Default S3 plot method for decon objects (derived from `deconvolve()`)}
\usage{
\method{plot}{decon}(x, bw = TRUE, ...)
}
\arguments{
\item{x}{decon object as generated by deconvolve}

\item{bw}{logical argument indicating whether the plot
should be in black and white or colour}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
This function sets up the default plotting method
for outputs from deconvolve function
}
