% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.process.R
\name{plot.process}
\alias{plot.process}
\title{Sets default plotting for process objects}
\usage{
\method{plot}{process}(x, plot_type = NULL, ...)
}
\arguments{
\item{x}{process object as generated by process}

\item{plot_type}{defaults to both plots. Can specify 'mass_loss' or 'rate_loss' curves by themselves.}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
This function sets up the default plotting method
for outputs from process function
}
