% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.process.R
\name{plot.process}
\alias{plot.process}
\title{Default S3 plot method for process objects (derived from `process()`)}
\usage{
\method{plot}{process}(x, plot_type = NULL, cex = 1, ...)
}
\arguments{
\item{x}{process object as generated by process}

\item{plot_type}{defaults to both plots. Can specify 'mass' or
'rate' curves by themselves.}

\item{cex}{size of plots features}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
This function sets up the default plotting method
for outputs from process function
}
