% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Processes Thermogravimetric Data}
\usage{
process(data, temp_col, massloss_col, massinit_value, temp_type = "C")
}
\arguments{
\item{data}{dataframe}

\item{temp_col}{column name containing temperature values}

\item{massloss_col}{column name containing mass loss values in grams}

\item{massinit_value}{numeric value of initial sample mass in grams}

\item{temp_type}{specify units of temperature, default = Celsius. Specify 'K' if in Kelvin}
}
\value{
decon list containing amended dataframe, bounds, model output, mass fractions
}
\description{
This function processes thermogravimetric data by calculating
the derivative of mass loss
}
\examples{
data(juncus)
tmp <- process(juncus, 'temp_C', 'mass_loss', 16.85)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
