% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_data.R
\name{rate_data}
\alias{rate_data}
\title{Accessor function to extract processed dataframe}
\usage{
rate_data(object)
}
\arguments{
\item{object}{a process or deconvolve object}
}
\value{
Dataframe of the object
}
\description{
Accessor function to extract processed dataframe
}
\examples{
data(juncus)
tmp <- process(juncus, init_mass = 18.96,
               temp = 'temp_C', mass_loss = 'mass_loss')
rate_data(tmp)

}
\keyword{deconvolution}
\keyword{fraser-suzuki}
\keyword{thermogravimetry}
