% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_quantiles.R
\name{wt_component}
\alias{wt_component}
\title{Calculate weight single component}
\usage{
wt_component(j, param_vec, lower_temp, upper_temp)
}
\arguments{
\item{j}{component}

\item{param_vec}{vector of parameters}

\item{lower_temp}{lower temperature bound}

\item{upper_temp}{upper temperature bound}
}
\value{
weight of component
}
\description{
Calculate weight single component
}
